# openapi.yaml
basePath: /
consumes:
- application/json
definitions:
...
  Document:
    properties:
      created_at:
        format: date-time
        type: string
      file_size:
        type: integer
      filename:
        type: string
      id:
        type: integer
      mime_type:
        type: string
      object_id:
        type: integer
      original_filename:
        type: string
      uploaded_by:
        type: string
    type: object
  Error:
    properties:
      error:
        description: Felmeddelande
        type: string
    type: object
...
  Object:
    properties:
      created_at:
        format: date-time
        type: string
      data:
        description: Fältvärden (nyckel = fältnamn)
        type: object
      id:
        type: integer
      id_full:
        description: Unikt sammansatt ID (t.ex. PROD-1.v1)
        type: string
      main_id:
        type: string
      object_type:
        type: string
      object_type_id:
        type: integer
      status:
        type: string
      updated_at:
        format: date-time
        type: string
      version:
        type: string
    type: object
...
info:
  contact:
    name: Ondfrukt
    url: https://github.com/ondfrukt/byggdelsuppdelning---demo
  description: REST API för Byggdelsuppdelning – ett PLM-system (Product Lifecycle
    Management) för hantering av byggdelar, objekttyper, relationer, dokument, klassificering
    och ändringshantering. Alla endpoints börjar med `/api/`.
  title: Byggdelsuppdelning API
  version: 1.0.0
paths:
...
  /api/objects:
    get:
      parameters:
      - description: Filtrera på objekttypens namn
        in: query
        name: type
        required: false
        type: string
      - description: Fritextsökning i fältvärden
        in: query
        name: search
        required: false
        type: string
      - default: false
        description: Returnera minimalt schema (ej fältdata)
        in: query
        name: minimal
        required: false
        type: boolean
      - description: Sidnummer för paginering
        in: query
        name: page
        required: false
        type: integer
      - description: Antal objekt per sida
        in: query
        name: per_page
        required: false
        type: integer
      - in: query
        name: sort_field
        required: false
        type: string
      - default: asc
        enum:
        - asc
        - desc
        in: query
        name: sort_direction
        required: false
        type: string
      - description: JSON-sträng med kolumnfiltrar
        in: query
        name: column_filters
        required: false
        type: string
      responses:
        '200':
          description: Lista med objekt (ev. paginerad)
          schema:
            items:
              $ref: '#/definitions/Object'
            type: array
        '500':
          description: Serverfel
          schema:
            $ref: '#/definitions/Error'
      summary: List all objects with optional filtering and optional pagination.
      tags:
      - Objects
    post:
      consumes:
      - application/json
      parameters:
      - in: body
        name: body
        required: true
        schema:
          properties:
            data:
              description: Fältvärden (nyckel = fältnamn)
              type: object
            main_id:
              description: Manuellt ID (annars auto-genererat)
              type: string
            object_type_id:
              description: Objekttypens ID
              type: integer
            status:
              description: Status
              type: string
            version:
              description: Versionstext (default v1)
              type: string
          required:
          - object_type_id
          type: object
      responses:
        '201':
          description: Objekt skapat
          schema:
            $ref: '#/definitions/Object'
        '400':
          description: Valideringsfel
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Serverfel
          schema:
            $ref: '#/definitions/Error'
      summary: Create a new object with dynamic data
      tags:
      - Objects
  /api/objects/documents/{doc_id}:
    delete:
      parameters:
      - description: Dokumentets ID
        in: path
        name: doc_id
        required: true
        type: integer
      responses:
        '200':
          description: Dokumentet borttaget
          schema:
            properties:
              message:
                type: string
            type: object
        '404':
          description: Dokumentet hittades inte
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Serverfel
          schema:
            $ref: '#/definitions/Error'
      summary: Delete a document
      tags:
      - Documents
  /api/objects/documents/{doc_id}/download:
    get:
      parameters:
      - description: Dokumentets ID
        in: path
        name: doc_id
        required: true
        type: integer
      - description: Tvinga nedladdning (annars inline för PDF)
        in: query
        name: download
        required: false
        type: boolean
      - description: Tvinga inline-visning
        in: query
        name: inline
        required: false
        type: boolean
      produces:
      - application/octet-stream
      - application/pdf
      responses:
        '200':
          description: Filinnehållet
        '404':
          description: Dokument eller fil hittades inte
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Serverfel
          schema:
            $ref: '#/definitions/Error'
      summary: Download a document
      tags:
      - Documents
  /api/objects/documents/{doc_id}/preview-meta:
    get:
      parameters:
      - description: Dokumentets ID (måste vara PDF)
        in: path
        name: doc_id
        required: true
        type: integer
      responses:
        '200':
          description: Sidmått för PDF
          schema:
            properties:
              document_id:
                type: integer
              orientation:
                enum:
                - portrait
                - landscape
                - square
                type: string
              page_height:
                type: number
              page_ratio:
                type: number
              page_width:
                type: number
            type: object
        '400':
          description: Dokumentet är inte en PDF
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: Dokument eller fil hittades inte
          schema:
            $ref: '#/definitions/Error'
        '422':
          description: PDF saknar sidor
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Serverfel
          schema:
            $ref: '#/definitions/Error'
      summary: Return first-page dimensions for PDF preview sizing.
      tags:
      - Documents
...
  /api/objects/{id}/documents:
    get:
      parameters:
      - description: Objektets ID (måste vara av typen FileObject)
        in: path
        name: id
        required: true
        type: integer
      responses:
        '200':
          description: Lista med dokument
          schema:
            items:
              $ref: '#/definitions/Document'
            type: array
        '404':
          description: Objektet hittades inte
          schema:
            $ref: '#/definitions/Error'
        '422':
          description: Objektet är inte av typen FileObject
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Serverfel
          schema:
            $ref: '#/definitions/Error'
      summary: List all documents for an object
      tags:
      - Documents
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Objektets ID (måste vara av typen FileObject)
        in: path
        name: id
        required: true
        type: integer
      - description: 'Tillåtna format: xls, xlsx, doc, docx, pdf, dwg, dxf, rvt, png,
          jpg, gif, bmp, webp, tif, tiff'
        in: formData
        name: file
        required: true
        type: file
      - description: Uppladdad av (valfri text)
        in: formData
        name: uploaded_by
        required: false
        type: string
      responses:
        '201':
          description: Dokument uppladdad
          schema:
            $ref: '#/definitions/Document'
        '400':
          description: Valideringsfel (filtyp, storlek)
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: Objektet hittades inte
          schema:
            $ref: '#/definitions/Error'
        '422':
          description: Objektet är inte av typen FileObject
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Serverfel
          schema:
            $ref: '#/definitions/Error'
      summary: Upload a document for an object
      tags:
      - Documents
...
produces:
- application/json
schemes:
- https
- http
securityDefinitions: {}
swagger: '2.0'
...